Refactor KubernetesLifecycle
