+++ server/src/main/java/org/opentosca/toscana/plugins/kubernetes/lifecycle/LifecycleHandler.java
@@ -1,0 +1,13 @@
+package org.opentosca.toscana.plugins.kubernetes.lifecycle;
+
+import org.slf4j.Logger;
+
+abstract class LifecycleHandler {
+    protected KubernetesLifecycle lifecycle;
+    protected Logger logger;
+
+    LifecycleHandler(KubernetesLifecycle lifecycle) {
+        this.lifecycle = lifecycle;
+        logger = lifecycle.getContext().getLogger(getClass());
+    }
+}
