+++ server/src/main/java/org/opentosca/toscana/plugins/kubernetes/model/kuberesource/ResourceService.java
@@ -1,7 +1,10 @@
-package org.opentosca.toscana.plugins.kubernetes.model;
+package org.opentosca.toscana.plugins.kubernetes.model.kuberesource;
 
 import java.util.List;
 import java.util.stream.Collectors;
+
+import org.opentosca.toscana.plugins.kubernetes.model.transform.Pod;
+import org.opentosca.toscana.plugins.kubernetes.model.transform.Port;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import io.fabric8.kubernetes.api.model.Service;
