+++ server/src/main/java/org/opentosca/toscana/plugins/kubernetes/model/transform/Port.java
@@ -1,4 +1,4 @@
-package org.opentosca.toscana.plugins.kubernetes.model;
+package org.opentosca.toscana.plugins.kubernetes.model.transform;
 
 import io.fabric8.kubernetes.api.model.ContainerPort;
 import io.fabric8.kubernetes.api.model.ContainerPortBuilder;
