+++ server/src/main/java/org/opentosca/toscana/plugins/kubernetes/model/transform/RelationshipGraph.java
@@ -1,4 +1,4 @@
-package org.opentosca.toscana.plugins.kubernetes.model;
+package org.opentosca.toscana.plugins.kubernetes.model.transform;
 
 import java.util.Set;
